agent-linux: send credentials regardless of site/realm
Another approach would be to subclass override get_basic_credentials(),
or to just parse $notify_url for the host & hardcode the realm.

That would be even better in some ways. However, the current approach
also gets rid of the initial unauthenticated request, which otherwise
would have doubled the number of requests...