ui-web: XHR replace only <tbody> again
(But keep XHR response as-is, because I still want to re-do the template
structure anyway.)

When replacing a whole <table>, Mozilla 1.7 noticeably flickers. This
doesn't happen when changing just the <tbody>, which is why we used to
do that in the first place.

This solution seems to satisfy both Mozilla *and* IE 6.