lib-php: we should only accept parameters as lists, not mappings
Even if we already expose the PHP method names as an API, I don't *yet*
want to allow this for parameter names.

More specifically, I don't want to allow calling the same method in two
different ways (list and mapping).

Still unsure about whether it's better to refuse mappings entirely, or
to pass them as a single parameter to the method.