lib-php: add several frontend-oriented RPC functions
Until now, all frontends had to be running on the same machine as the
server (to have DB access) and on the same machine as each other (to
share configuration such as expiry times, not to mention .plan files).

This is mostly important for in.rwho-finger, which has to be running
wherever the zone apex domain points... so when it was moved to star,
the entirety of Rwho had to travel with it.

But now that we have a proper RPC interface, frontends can begin using
it instead of directly reading from the database.