lib-php: support retrieving utmp entries by short hostname
The purpose of this is to allow `finger @ember@nullroute.lt` without
requiring complex changes to the frontends themselves (e.g. retrying
with .domain appended -- which would require the domain to be
configurable -- or using "LIKE foo.%" queries).

An alternative approach would be to store the short hostname in the
`host` column and introduce `raw_host`, similar to how we already have
the `raw_user` field for SSSD "@domain" stripping. I will likely
implement that as well.